delete alert

+++ src/components/quiz-management/QuizManagementTable.jsx
@@ -1,23 +1,33 @@
 import React from "react";
-import { FaEdit, FaTrash } from "react-icons/fa";
+//import { FaEdit, FaTrash } from "react-icons/fa";
+import { LiaEdit } from "react-icons/lia";
+import { LiaTrashAltSolid } from "react-icons/lia";
 import { formatToDateTimeString } from "../../utilities";
 import { deleteQuiz } from "../../api/apiService";
 import { useNavigate } from "react-router-dom";
-import Swal from 'sweetalert2'
+import Swal from "sweetalert2";
 
 const QuizManagementTable = ({ data = [], onDelete = () => {} }) => {
   const handleDelete = async (quizId) => {
+    const confirm = await Swal.fire({
+      title: "Are you sure to delete?",
+      showDenyButton: true,
+      confirmButtonText: "Delete",
+      denyButtonText: `Cancel`,
+    });
+
+    if (confirm.isDenied || confirm.isDismissed) {
+      return;
+    }
+
     try {
       await deleteQuiz(quizId);
-      // setData((prevData) => prevData.filter((quiz) => quiz.quizId !== quizId));
-      // alert("Quiz deleted successfully.");
       await Swal.fire({
         title: "Quiz Deleted Successfully!",
         icon: "success",
-        draggable: true
+        draggable: true,
       });
       onDelete(quizId);
-
     } catch (error) {
       console.error("Failed to delete quiz:", error);
       Swal.fire("Error", "Failed to delete quiz.", "error");
@@ -47,13 +57,18 @@ const QuizManagementTable = ({ data = [], onDelete = () => {} }) => {
               <td>{quiz.duration} min</td>
               <td>{formatToDateTimeString(quiz.createdAt)}</td>
               <td>
-                <FaEdit
-                  style={{ cursor: "pointer", marginRight: "10px" }}
+                <LiaEdit
+                  style={{
+                    color : "black" ,
+                    cursor: "pointer",
+                    marginRight: "10px",
+                    fontSize: "1.5rem",
+                  }}
                   onClick={() => handleEdit(quiz.quizId)}
                 />
-                <FaTrash
+                <LiaTrashAltSolid
                   className="ms-2"
-                  style={{ cursor: "pointer" }}
+                  style={{ color : "red" ,cursor: "pointer", fontSize: "1.5rem" }}
                   onClick={() => handleDelete(quiz.quizId)}
                 />
               </td>
